feat: suite du system site_v2
