+++ plugin.video.vstream/resources/site_v2/serie/result_serie.py
@@ -1,0 +1,31 @@
+# -*- coding: utf-8 -*-
+# vStream https://github.com/Kodi-vStream/venom-xbmc-addons
+from __future__ import annotations
+
+from typing import TYPE_CHECKING
+
+from bs4 import BeautifulSoup
+
+from resources.site_v2.result import Result
+
+if TYPE_CHECKING:
+    from resources.site_v2.site import Site
+
+
+class ResultSerie(Result):
+
+    def __init__(self, site: Site):
+        super().__init__(site)
+
+    def __str__(self) -> str:
+        return "<serie> " + super().__str__()
+
+    @staticmethod
+    def from_result(result: Result) -> ResultSerie:
+        serie_result = ResultSerie(result._site)
+        serie_result.update(result)
+        return serie_result
+
+    def get_content(self) -> BeautifulSoup:
+        self._fetch_content_if_none()
+        return self._content
